+++ quic/s2n-quic-transport/src/stream/send_stream/tests.rs
@@ -360,7 +360,7 @@ fn can_not_enqueue_data_if_max_buffer_size_has_been_reached() {
             // as long as we are below the flow control window. We subtract 2
             // in the second instruction so that even the the third instruction
             // would not cause us a rejection based on the flow control size.
-            Instruction::EnqueueData(VarInt::from_u32(0), MAX_BUFFER_SIZE as usize - 1, true),
+            Instruction::EnqueueData(VarInt::from_u32(0), MAX_BUFFER_SIZE - 1, true),
             Instruction::EnqueueData(
                 max_buffer_size_varint - 1,
                 RECEIVE_WINDOW as usize - MAX_BUFFER_SIZE - 2,
@@ -488,7 +488,7 @@ fn multiple_stream_frames_are_sent_in_a_packet() {
             // as long as we are below the flow control window. We subtract 2
             // in the second instruction so that even the the third instruction
             // would not cause us a rejection based on the flow control size.
-            Instruction::EnqueueData(VarInt::from_u32(0), MAX_BUFFER_SIZE as usize - 1, true),
+            Instruction::EnqueueData(VarInt::from_u32(0), MAX_BUFFER_SIZE - 1, true),
             Instruction::EnqueueData(
                 max_buffer_size_varint - 1,
                 RECEIVE_WINDOW as usize - MAX_BUFFER_SIZE - 2,
